requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.78
botocore==1.31.78
Bottleneck==1.3.5
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.3
DataProperty==1.0.1
datasets==2.14.5
dill==0.3.7
dyNET38==2.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-events==0.9.1
filelock==3.12.4
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==1.1.1
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==4.0.0
kiwisolver==1.4.4
lm-eval==0.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mbstrdecoder==1.1.3
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.2
multidict==6.0.2
multiprocess==0.70.15
munkres==1.1.4
nagisa==0.2.9
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
omegaconf==2.3.0
openai==0.28.1
openpyxl==3.0.10
packaging==23.1
pandas==2.1.1
pandas-stubs==1.5.3.230203
pathvalidate==3.2.0
peft==0.5.0
Pillow==10.0.1
pip==23.3
plotly==5.9.0
portalocker==2.8.2
psutil==5.9.5
pyarrow==13.0.0
pybind11==2.11.1
pycountry==22.3.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pynvim==0.4.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytablewriter==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rouge-score==0.1.2
s3transfer==0.7.0
sacrebleu==1.5.0
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.3
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sqlitedict==2.1.0
starlette==0.27.0
sympy==1.12
tabledata==1.3.3
tcolorpy==0.1.4
tenacity==8.2.2
threadpoolctl==2.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.65.0
tqdm-multiprocess==0.0.11
transformers==4.34.0
typepy==1.3.2
types-pytz==2022.4.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wheel==0.41.2
xxhash==3.4.1
yarl==1.8.1
zipp==3.11.0
zstandard==0.21.0
